Move `ChartOutput` to common module
